+++ tests/test_package.py
@@ -2,4 +2,4 @@ import hyperclass
 
 
 def test_version():
-    assert hyperclass.__version__ == "0.1.0"
+    assert hyperclass.__version__ == "0.1.1"